button component for code reuse
