+++ src/components/Button/index.js
@@ -1,26 +1,4 @@
-import React from 'react';
-import {TouchableOpacity, StyleSheet} from 'react-native';
-import Icon from 'react-native-vector-icons/AntDesign';
+import AddBtn from './AddBtn';
+import TextBtn from './TextBtn';
 
-const addBtn = props => {
-  return (
-    <TouchableOpacity
-      style={styles.button}
-      onPress={() => props.onButtonClick()}>
-      <Icon name="plus" size={20} color="#FFFFFF" />
-    </TouchableOpacity>
-  );
-};
-
-const styles = StyleSheet.create({
-  button: {
-    height: 56,
-    width: 56,
-    backgroundColor: '#F9C201',
-    borderRadius: 15,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
-
-export default addBtn;
+export {AddBtn, TextBtn};
