+++ src/components/index.js
@@ -1,7 +1,5 @@
-import Button from './Button';
-import SaveButton from './SaveButton';
 import TrackHistory from './TrackHistory';
 import TextBox from './TextBox';
 import SwitchToggle from './SwitchToggle';
 
-export {Button, SaveButton, TrackHistory, TextBox, SwitchToggle};
+export {TrackHistory, TextBox, SwitchToggle};
